+++ src/components/EditSections/EditReadingRoomAccess/EditReadingRoomAccess.js
@@ -1,17 +1,20 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { noop } from 'lodash';
+import { cloneDeep, noop, orderBy } from 'lodash';
 
 import {
   Accordion,
   Badge,
   Headline,
+  Loading,
   MultiColumnList,
 } from '@folio/stripes/components';
 
-import { rraColumns } from './constants';
+import { rraColumns, DEFAULT_SORT_ORDER } from './constants';
 import { getFormatter } from './getFormatter';
+import { sortTypes } from '../../../constants';
+import { getReadingRoomSortedData } from '../../util/util';
 
 const EditReadingRoomAccess = ({
   expanded,
@@ -30,6 +33,13 @@ const EditReadingRoomAccess = ({
     [rraColumns.ACCESS]: '15%',
     [rraColumns.READING_ROOM_NAME]: '25%',
   };
+  const sortInitialState = {
+    data: [],
+    order: DEFAULT_SORT_ORDER,
+    direction: sortTypes.ASC,
+  };
+  const [sortedRecordsDetails, setSortedRecordsDetails] = useState(sortInitialState);
+  const [sortedRecordsIsLoading, setSortedRecordsIsLoading] = useState(false);
 
   useEffect(() => {
     const unregisterReadingRoomAccessList = form.registerField('readingRoomsAccessList', noop, { initialValue: [] });
@@ -39,6 +49,35 @@ const EditReadingRoomAccess = ({
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
+  useEffect(() => {
+    setSortedRecordsDetails(prev => ({
+      ...prev,
+      data: orderBy(formData, [data => data[prev.order]?.toLowerCase()], prev.direction)
+    }));
+  }, [formData, form]);
+
+  const onSort = (e, meta) => {
+    setSortedRecordsIsLoading(true);
+    const recordDetails = cloneDeep(sortedRecordsDetails);
+    const updatedRecords = form.getFieldState('readingRoomsAccessList').value;
+
+    updatedRecords?.forEach(updatedRecord => {
+      const index = recordDetails?.data?.findIndex(record => record.readingRoomId === updatedRecord.readingRoomId);
+      if (index !== -1) {
+        recordDetails.data[index] = updatedRecord;
+      }
+    });
+
+    setSortedRecordsDetails(
+      getReadingRoomSortedData(e, meta, recordDetails)
+    );
+  // setState callback, use Transition
+  };
+
+  useEffect(() => {
+    setSortedRecordsIsLoading(false);
+  }, [sortedRecordsDetails]);
+
   return (
     <Accordion
       open={expanded}
@@ -47,14 +86,19 @@ const EditReadingRoomAccess = ({
       label={<Headline size="large" tag="h3"><FormattedMessage id="ui-users.readingRoom.readingRoomAccess" /></Headline>}
       displayWhenClosed={<Badge>{formData.length}</Badge>}
     >
+      { sortedRecordsIsLoading ? <Loading /> :
       <MultiColumnList
         striped
-        contentData={formData}
+        contentData={sortedRecordsDetails?.data}
         columnMapping={columnMapping}
         visibleColumns={visibleColumns}
         formatter={getFormatter(form)}
         columnWidths={columnWidths}
-      />
+        sortOrder={sortedRecordsDetails.order}
+        sortDirection={`${sortedRecordsDetails.direction}ending`}
+        onHeaderClick={onSort}
+        sortedColumn={sortedRecordsDetails.order}
+      /> }
     </Accordion>
   );
 };